requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
PyPDF2>=3.0.0
requests>=2.28.0
python-dotenv>=0.19.0
reportlab>=4.0.0
python-docx>=1.0.0 